refactor: 타입 선언 한 파일로 합침
