+++ src/apis/cart.ts
@@ -1,21 +1,20 @@
-import { CartItem } from '../types/cart';
+import type { CartItem } from '../types/cart';
 import { waitFor, WaitForOptions } from '../utils/waitFor';
 import { authFetchQuery } from './api';
-import { FetchQueryRes } from './api.type';
-
-export type FetchCartRes = CartItem[];
+import type {
+  AddCartDataReq,
+  AddCartDataRes,
+  DeleteCartItemRes,
+  FetchCartRes,
+  FetchQueryRes,
+  UpdateCartItemRes,
+} from './api.type';
 
 export const fetchCart = (options?: WaitForOptions<FetchCartRes>) => {
   const promise = authFetchQuery.get<FetchCartRes>(`/cart-items`);
 
   return waitFor(promise, options);
 };
-
-export interface AddCartDataReq {
-  productId: number;
-}
-
-interface AddCartDataRes {}
 
 export const addToCart: (
   payload: AddCartDataReq
@@ -25,13 +24,6 @@ export const addToCart: (
   });
 };
 
-export interface UpdateCartItemReq {
-  cartId: number;
-  quantity: number;
-}
-
-interface UpdateCartItemRes {}
-
 export const updateCartItem: (
   payload: UpdateCartItemReq
 ) => FetchQueryRes<UpdateCartItemRes> = ({ cartId, quantity }) => {
@@ -40,10 +32,13 @@ export const updateCartItem: (
   });
 };
 
-interface DeleteCartItemRes {}
-
 export const deleteCartItem: (
   id: CartItem['id']
 ) => FetchQueryRes<DeleteCartItemRes> = (id) => {
   return authFetchQuery.delete<DeleteCartItemRes>(`/cart-items/${id}`);
 };
+
+export interface UpdateCartItemReq {
+  cartId: number;
+  quantity: number;
+}
