+++ src/apis/coupons.ts
@@ -1,10 +1,5 @@
-import { Coupon, SpecificCoupon } from '../types/coupon';
 import { authFetchQuery } from './api';
-
-interface FetchCouponsRes {
-  allCoupons: Coupon[];
-  specificCoupons: SpecificCoupon[];
-}
+import type { FetchCouponsRes } from './api.type';
 
 export const fetchCoupons = async () => {
   return authFetchQuery.get<FetchCouponsRes>('/coupons');
