+++ src/apis/products.ts
@@ -1,8 +1,5 @@
-import { Product } from '../types/products';
+import type { FetchProductDataRes, FetchQueryRes } from './api.type';
 import { fetchQuery } from './api';
-import { FetchQueryRes } from './api.type';
-
-type FetchProductDataRes = Product[];
 
 export const fetchProductData =
   async (): FetchQueryRes<FetchProductDataRes> => {
